requirements.txt
PyAudio
numpy
websocket-client
python-socketio[client]